Greatly improved the overall synchronization (in response to issue #16) It appears that now the reactor is able to find a solution or otherwise trigger a relax.
Improved archiving (issue #14). Still it looks like I have a lot of dangling tokens ... maybe it is worth doing a loop until I have no more things to archive

Added a new agent with the light reactor using a slight modification of lightswitch model. This one is sluggish but it demonstrates that you can use a model for simulation